V 1.11.15 Of NukaCoin! FIRST PUBLIC RELEASE
We have brought back NUKACOIN! This is the first release of the new and revived NukaCoin, built on the NXT blockchain. This is a very early version, we have not changed any assets or logos on this one, though most bugs have been crushed. We are still working on getting peers and nodes working properly. As this is a pre-issued currency, there is no way for you to get the currency on your own. To get 1000 coins for yourself, join our Discord and message BrandonAxtmann! We will be creating a faucet at a later time to help fix this.